+++ packages/bezier-figma-plugin/src/ui/components/ExtractSuccess.tsx
@@ -11,7 +11,6 @@ import {
   ButtonStyleVariant,
   StackItem,
   Text,
-  Typography,
   VStack,
 } from '@channel.io/bezier-react'
 
@@ -27,7 +26,7 @@ function ExtractSuccess() {
   return (
     <VStack align="center" justify="center" spacing={2}>
       <StackItem>
-        <Text typo={Typography.Size18}>아이콘 추출 성공!</Text>
+        <Text typo="18">아이콘 추출 성공!</Text>
       </StackItem>
       <StackItem>
         { /* @ts-ignore */ }