+++ packages/bezier-react/.storybook/preview.ts
@@ -1,5 +1,7 @@
 import type { Preview } from "@storybook/react"
 
+import '~/src/styles/index.scss'
+
 import { WithFoundationProvider } from './WithFoundationProvider'
 
 const preview: Preview = {