+++ packages/bezier-react/src/components/Button/Button.test.tsx
@@ -2,9 +2,7 @@ import React from 'react'
 
 import {
   LightFoundation,
-  LineHeightAbsoluteNumber,
   RoundAbsoluteNumber,
-  TypoAbsoluteNumber,
 } from '~/src/foundation'
 
 import { render } from '~/src/utils/test'
@@ -173,10 +171,6 @@ describe('Button Test >', () => {
         // eslint-disable-next-line max-len
         expect(defaultButton).toHaveStyle(`padding: 0 ${defaultButtonPadding}px 0 ${defaultButtonPadding}px;`)
 
-        // Typograpy.Size14
-        expect(defaultButtonText).toHaveStyle(`font-size: ${TypoAbsoluteNumber.Typo14}rem;`)
-        expect(defaultButtonText).toHaveStyle(`line-height: ${LineHeightAbsoluteNumber.Lh18}rem;`)
-
         // Text padding value by ButtonSize
         expect(defaultButtonText).toHaveStyle(`padding: 0 ${TEXT_PADDING_VALUE[ButtonSize.M]}px;`)
 
@@ -192,10 +186,6 @@ describe('Button Test >', () => {
         expect(xsButton).toHaveStyle('min-width: 20px;')
         expect(xsButton).toHaveStyle('height: 20px;')
         expect(xsButton).toHaveStyle(`padding: 0 ${xsDefaultButtonPadding}px 0 ${xsDefaultButtonPadding}px;`)
-
-        // Typograpy.Size13
-        expect(xsButtonText).toHaveStyle(`font-size: ${TypoAbsoluteNumber.Typo13}rem;`)
-        expect(xsButtonText).toHaveStyle(`line-height: ${LineHeightAbsoluteNumber.Lh18}rem;`)
 
         // Text padding value by ButtonSize
         expect(xsButtonText).toHaveStyle(`padding: 0 ${TEXT_PADDING_VALUE[ButtonSize.XS]}px;`)
@@ -225,10 +215,6 @@ describe('Button Test >', () => {
         expect(sButton).toHaveStyle('min-width: 24px;')
         expect(sButton).toHaveStyle('height: 24px;')
         expect(sButton).toHaveStyle(`padding: 0 ${sDefaultButtonPadding}px 0 ${sDefaultButtonPadding}px;`)
-
-        // Typograpy.Size13
-        expect(sButtonText).toHaveStyle(`font-size: ${TypoAbsoluteNumber.Typo13}rem;`)
-        expect(sButtonText).toHaveStyle(`line-height: ${LineHeightAbsoluteNumber.Lh18}rem;`)
 
         // Text padding value by ButtonSize
         expect(sButtonText).toHaveStyle(`padding: 0 ${TEXT_PADDING_VALUE[ButtonSize.S]}px;`)
@@ -260,10 +246,6 @@ describe('Button Test >', () => {
         // eslint-disable-next-line max-len
         expect(mButton).toHaveStyle(`padding: 0 ${mDefaultButtonPadding}px 0 ${mDefaultButtonPadding}px;`)
 
-        // Typograpy.Size14
-        expect(mButtonText).toHaveStyle(`font-size: ${TypoAbsoluteNumber.Typo14}rem;`)
-        expect(mButtonText).toHaveStyle(`line-height: ${LineHeightAbsoluteNumber.Lh18}rem;`)
-
         // Text padding value by ButtonSize
         expect(mButtonText).toHaveStyle(`padding: 0 ${TEXT_PADDING_VALUE[ButtonSize.M]}px;`)
 
@@ -293,10 +275,6 @@ describe('Button Test >', () => {
         expect(lButton).toHaveStyle('height: 44px;')
         expect(lButton).toHaveStyle(`padding: 0 ${lDefaultButtonPadding}px 0 ${lDefaultButtonPadding}px;`)
 
-        // Typography.Size15
-        expect(lButtonText).toHaveStyle(`font-size: ${TypoAbsoluteNumber.Typo15}rem;`)
-        expect(lButtonText).toHaveStyle(`line-height: ${LineHeightAbsoluteNumber.Lh20}rem;`)
-
         // Text padding value by ButtonSize
         expect(lButtonText).toHaveStyle(`padding: 0 ${TEXT_PADDING_VALUE[ButtonSize.L]}px;`)
 
@@ -325,10 +303,6 @@ describe('Button Test >', () => {
         expect(xlButton).toHaveStyle('min-width: 54px;')
         expect(xlButton).toHaveStyle('height: 54px;')
         expect(xlButton).toHaveStyle(`padding: 0 ${xlDefaultButtonPadding}px 0 ${xlDefaultButtonPadding}px;`)
-
-        // Typography.Size18
-        expect(xlButtonText).toHaveStyle(`font-size: ${TypoAbsoluteNumber.Typo18}rem;`)
-        expect(xlButtonText).toHaveStyle(`line-height: ${LineHeightAbsoluteNumber.Lh24}rem;`)
 
         // Text padding value by ButtonSize
         expect(xlButtonText).toHaveStyle(`padding: 0 ${TEXT_PADDING_VALUE[ButtonSize.XL]}px;`)