+++ packages/bezier-react/src/components/Forms/SegmentedControl/SegmentedControl.tsx
@@ -8,8 +8,7 @@ import * as RadioGroupPrimitive from '@radix-ui/react-radio-group'
 import * as TabsPrimitive from '@radix-ui/react-tabs'
 import classNames from 'classnames'
 
-import { px } from '~/src/utils/style'
-import { isNumber } from '~/src/utils/type'
+import { cssDimension } from '~/src/utils/style'
 
 import { Divider } from '~/src/components/Divider'
 import useFormFieldProps from '~/src/components/Forms/useFormFieldProps'
@@ -63,7 +62,7 @@ function SegmentedControlItemListImpl<
 
   const style = useMemo(() => ({
     ...styleProp,
-    '--b-segmented-control-width': isNumber(width) ? px(width) : width,
+    '--b-segmented-control-width': cssDimension(width),
   } as React.CSSProperties), [
     styleProp,
     width,