+++ packages/bezier-react/src/components/TagBadge/TagBadgeCommon/TagBadgeText/TagBadgeText.types.ts
@@ -2,15 +2,16 @@ import {
   type BezierComponentProps,
   type ChildrenProps,
 } from '~/src/types/ComponentProps'
-import { type InjectedInterpolation } from '~/src/types/Foundation'
+
+import { type TextProps } from '~/src/components/Text'
 
 interface TagBadgeTextOptions {
-  typo: InjectedInterpolation
   horizontalPadding: number
 }
 interface TagBadgeTextProps extends
   Omit<BezierComponentProps, 'style' | 'className' | 'interpolation'>,
   ChildrenProps,
+  Pick<TextProps, 'typo'>,
   TagBadgeTextOptions {}
 
 export default TagBadgeTextProps