+++ packages/bezier-react/src/components/Text/Text.stories.tsx
@@ -3,81 +3,50 @@ import React from 'react'
 import {
   type Meta,
   type StoryFn,
+  type StoryObj,
 } from '@storybook/react'
 
-import Text from './Text'
-import type TextProps from './Text.types'
+import { Text } from './Text'
+import { type TextProps } from './Text.types'
 
 const meta: Meta<typeof Text> = {
   component: Text,
-  argTypes: {
-    /**
-     * FIXME:
-     * storybook controls 에서 styled-components 의 css ReturnType 를
-     * select options 의 entity 로 지원하지 않는 듯 함.
-     * 임시로 typo 를 controls 에서 제거하였음.
-     */
-    typo: { table: { disable: true } },
-    as: {
-      control: {
-        type: 'select',
-      },
-      options: [
-        undefined,
-        'h1',
-        'h2',
-        'button',
-      ],
-    },
-
-    style: { control: 'object' },
-  },
 }
-export default meta
 
-const Template: StoryFn<TextProps> = ({ children, ...otherTextProps }) => (
-  <Text {...otherTextProps}>{ children }</Text>
+const Template: StoryFn<TextProps> = ({ children, ...rest }) => (
+  <Text {...rest}>
+    { children }
+  </Text>
 )
 
-export const Primary = {
+export const Primary: StoryObj<typeof Text> = {
   render: Template,
-
   args: {
-    as: undefined,
-    bold: false,
-    italic: false,
-    truncated: false,
-    style: { color: 'gray' },
-    children: 'hello',
-    marginTop: 0,
-    marginRight: 0,
-    marginBottom: 0,
-    marginLeft: 0,
-    marginVertical: 0,
-    marginHorizontal: 0,
-    marginAll: 0,
+    color: 'txt-black-darkest',
+    children: 'Hello, Channel!',
   },
 }
 
-const Truncated: StoryFn<TextProps & { width: string }> = ({
+const Truncated: StoryFn<TextProps> = ({
   children,
-  width,
-  ...otherTextProps
+  ...rest
 }) => (
-  <div style={{ width }}>
-    <Text {...otherTextProps}>{ children }</Text>
+  <div style={{
+    width: '100px',
+  }}
+  >
+    <Text {...rest}>{ children }</Text>
   </div>
 )
 
-export const Secondary = {
+export const Secondary: StoryObj<typeof Text> = {
   render: Truncated,
-
   args: {
-    width: '100px',
+    color: 'txt-black-darkest',
     truncated: true,
-    style: { color: 'gray' },
-    children: 'test truncated long text',
+    children: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
   },
-
   name: 'Usage (truncated)',
 }
+
+export default meta