+++ packages/bezier-react/src/foundation/Typography.stories.tsx
@@ -57,7 +57,7 @@ export const Overview: StoryFn<{}> = () => (
   <VStack spacing={16}>
     <StackItem>
       <Text
-        typo={Typography.Size11}
+        typo="11"
         color="txt-black-darkest"
       >
         11Aa한글 <b>11Aa한글</b>
@@ -65,7 +65,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size12}
+        typo="12"
         color="txt-black-darkest"
       >
         12Aa한글 <b>12Aa한글</b>
@@ -73,7 +73,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size13}
+        typo="13"
         color="txt-black-darkest"
       >
         13Aa한글 <b>13Aa한글</b>
@@ -81,7 +81,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size14}
+        typo="14"
         color="txt-black-darkest"
       >
         14Aa한글 <b>14Aa한글</b>
@@ -89,7 +89,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size15}
+        typo="15"
         color="txt-black-darkest"
       >
         15Aa한글 <b>15Aa한글</b>
@@ -97,7 +97,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size16}
+        typo="16"
         color="txt-black-darkest"
       >
         16Aa한글 <b>16Aa한글</b>
@@ -105,7 +105,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size17}
+        typo="17"
         color="txt-black-darkest"
       >
         17Aa한글 <b>17Aa한글</b>
@@ -113,7 +113,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size18}
+        typo="18"
         color="txt-black-darkest"
       >
         18Aa한글 <b>18Aa한글</b>
@@ -121,7 +121,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size22}
+        typo="22"
         color="txt-black-darkest"
       >
         22Aa한글 <b>22Aa한글</b>
@@ -129,7 +129,7 @@ export const Overview: StoryFn<{}> = () => (
     </StackItem>
     <StackItem>
       <Text
-        typo={Typography.Size24}
+        typo="24"
         color="txt-black-darkest"
       >
         24Aa한글 <b>24Aa한글</b>
@@ -140,7 +140,7 @@ export const Overview: StoryFn<{}> = () => (
 
 export const UsageText: StoryObj<{}> = {
   render: () => (
-    <Text as="p" typo={Typography.Size15} color="txt-black-darkest">
+    <Text as="p" typo="15" color="txt-black-darkest">
       Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent bibendum
       elit erat, eu euismod odio tincidunt vel. Nulla non consequat ligula.
       Nulla quis justo at enim sodales aliquet. Proin placerat varius elit, et
@@ -153,23 +153,4 @@ export const UsageText: StoryObj<{}> = {
   ),
 
   name: 'Usage (via Text)',
-}
-
-const TypographicCode = styled.code`
-  ${Typography.Size15}
-  color: ${({ foundation }) => foundation?.theme['txt-black-darkest']};
-`
-
-export const UsageInterpolation: StoryObj<{}> = {
-  render: () => (
-    <pre>
-      <TypographicCode>
-        { `function main() {
-    console.log('hello, world!');
-  }` }
-      </TypographicCode>
-    </pre>
-  ),
-
-  name: 'Usage (via interpolation)',
 }